Поправлены иниты. Вью стартового контроллера вынесен в отдельный класс.
