+++ UiLuxury_TeamProject/Scenes/MainScene/GameScene/GameViewController.swift
@@ -63,6 +63,7 @@ final class GameViewController: UIViewController {
         super.init(nibName: nil, bundle: nil)
     }
     
+    @available(*, unavailable)
     required init?(coder: NSCoder) {
         fatalError(GlobalConstants.fatalError)
     }
